+++ scripts/gen_changelog_html.js
@@ -1,0 +1,13 @@
+#!/usr/bin/env node
+
+const fs = require('fs');
+const marked = require('marked');
+
+const changelog = fs.readFileSync('CHANGELOG.md', 'utf8');
+
+console.log(marked.parse(changelog, (err, html) => {
+    if (err) {
+        console.error(err);
+        process.exit(1);
+    }
+}));