+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2025-01-20
+
 ### Added
 
 - Changelog dialog
@@ -421,7 +423,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial Release
 
-[unreleased]: https://github.com/Coteh/wordle-clone/compare/v1.3.5...HEAD
+[unreleased]: https://github.com/Coteh/wordle-clone/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/Coteh/wordle-clone/compare/v1.3.5...v1.4.0
 [1.3.5]: https://github.com/Coteh/wordle-clone/compare/v1.3.4...v1.3.5
 [1.3.4]: https://github.com/Coteh/wordle-clone/compare/v1.3.3...v1.3.4
 [1.3.3]: https://github.com/Coteh/wordle-clone/compare/v1.3.2...v1.3.3